Extract light environment audits
